+++ Changelog.md
@@ -12,6 +12,7 @@
 * Parsing `AlternateContent#ole_object`
 * Parsing `OOXMLShapeBodyProperties#number_columns`
 * Parsing `OOXMLShapeBodyProperties#space_columns`
+* Parsing `DocxParagraph#sdt`
 
 ### Refactor
 * Change interface of `XlsxDrawing#graphic_frame`